+++ lib/process_hub/utility/name.ex
@@ -19,6 +19,18 @@ defmodule ProcessHub.Utility.Name do
     |> String.to_atom()
   end
 
+  # TODO: add tests
+  # @spec cache(ProcessHub.hub_id()) :: atom()
+  # def cache(hub_id) do
+  #   concat_name([hub_id, "cache"], ".")
+  # end
+
+  # TODO: add tests
+  @spec worker_queue(ProcessHub.hub_id()) :: atom()
+  def worker_queue(hub_id) do
+    concat_name([hub_id, "worker_queue"], ".")
+  end
+
   @doc "The name of the main initializer process."
   @spec initializer(ProcessHub.hub_id()) :: atom()
   def initializer(hub_id) do