+++ test/integration_test.exs
@@ -18,66 +18,66 @@ defmodule Test.IntegrationTest do
     Test.Helper.Bootstrap.bootstrap(context)
   end
 
-  @tag hub_id: :pubsub_start_rem_test
-  @tag sync_strategy: :pubsub
-  @tag listed_hooks: [
-         {Hook.cluster_join(), :local},
-         {Hook.registry_pid_inserted(), :global},
-         {Hook.registry_pid_removed(), :global}
-       ]
-  test "pubsub children starting and removing", %{hub_id: hub_id} = context do
-    child_count = 100
-    child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
+  # @tag hub_id: :pubsub_start_rem_test
+  # @tag sync_strategy: :pubsub
+  # @tag listed_hooks: [
+  #        {Hook.cluster_join(), :local},
+  #        {Hook.registry_pid_inserted(), :global},
+  #        {Hook.registry_pid_removed(), :global}
+  #      ]
+  # test "pubsub children starting and removing", %{hub_id: hub_id} = context do
+  #   child_count = 100
+  #   child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
 
-    # TODO: this is failing sometimes
-    # Starts children on all nodes.
-    Common.sync_base_test(context, child_specs, :add, scope: :global)
+  #   # TODO: this is failing sometimes
+  #   # Starts children on all nodes.
+  #   Common.sync_base_test(context, child_specs, :add, scope: :global)
 
-    # Tests if all child_specs are used for starting children.
-    Common.validate_registry_length(context, child_specs)
+  #   # Tests if all child_specs are used for starting children.
+  #   Common.validate_registry_length(context, child_specs)
 
-    # Tests if all child_specs are started on all nodes.
-    Common.validate_started_children(context, child_specs)
+  #   # Tests if all child_specs are started on all nodes.
+  #   Common.validate_started_children(context, child_specs)
 
-    # Tests children adding and syncing.
-    Common.validate_sync(context)
+  #   # Tests children adding and syncing.
+  #   Common.validate_sync(context)
 
-    # Stops children on all nodes.
-    Common.sync_base_test(context, child_specs, :rem, scope: :global)
+  #   # Stops children on all nodes.
+  #   Common.sync_base_test(context, child_specs, :rem, scope: :global)
 
-    # Tests children removing and syncing.
-    Common.validate_sync(context)
-  end
+  #   # Tests children removing and syncing.
+  #   Common.validate_sync(context)
+  # end
 
-  @tag hub_id: :pubsub_interval_test
-  @tag sync_strategy: :pubsub
-  @tag listed_hooks: [
-         {Hook.cluster_join(), :local},
-         {Hook.registry_pid_inserted(), :global},
-         {Hook.registry_pid_removed(), :global}
-       ]
-  test "pubsub interval sync test", %{hub_id: hub_id} = context do
-    child_count = 100
-    child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
+  # @tag hub_id: :pubsub_interval_test
+  # @tag sync_strategy: :pubsub
+  # @tag listed_hooks: [
+  #        {Hook.cluster_join(), :local},
+  #        {Hook.registry_pid_inserted(), :global},
+  #        {Hook.registry_pid_removed(), :global}
+  #      ]
+  # test "pubsub interval sync test", %{hub_id: hub_id} = context do
+  #   child_count = 100
+  #   child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
 
-    # Locally start children without propagating to the rest of the cluster.
-    Common.periodic_sync_base(context, child_specs, :add)
+  #   # Locally start children without propagating to the rest of the cluster.
+  #   Common.periodic_sync_base(context, child_specs, :add)
 
-    # Manually trigger the periodic sync.
-    Common.trigger_periodc_sync(context, child_specs, :add)
+  #   # Manually trigger the periodic sync.
+  #   Common.trigger_periodc_sync(context, child_specs, :add)
 
-    # Test if data is synchronized in the cluster.
-    Common.validate_sync(context)
+  #   # Test if data is synchronized in the cluster.
+  #   Common.validate_sync(context)
 
-    # Locally stop children without propagating to the rest of the cluster.
-    Common.periodic_sync_base(context, child_specs, :rem)
+  #   # Locally stop children without propagating to the rest of the cluster.
+  #   Common.periodic_sync_base(context, child_specs, :rem)
 
-    # Manually trigger the periodic sync.
-    Common.trigger_periodc_sync(context, child_specs, :rem)
+  #   # Manually trigger the periodic sync.
+  #   Common.trigger_periodc_sync(context, child_specs, :rem)
 
-    # Test if data is synchronized in the cluster.
-    Common.validate_sync(context)
-  end
+  #   # Test if data is synchronized in the cluster.
+  #   Common.validate_sync(context)
+  # end
 
   # @tag hub_id: :gossip_start_rem_test
   # @tag sync_strategy: :gossip
@@ -103,6 +103,11 @@ defmodule Test.IntegrationTest do
   #   # Tests children adding and syncing.
   #   Common.validate_sync(context)
 
+  #   # Process.sleep(1000)
+  #   # Bag.all_messages()
+  #   # |> IO.inspect(label: "DEBUG ALL MESSAGES")
+
+  #   # TODO: this isfailing sometimes
   #   # Stops children on all nodes.
   #   Common.sync_base_test(context, child_specs, :rem, scope: :global)
 
@@ -110,37 +115,35 @@ defmodule Test.IntegrationTest do
   #   Common.validate_sync(context)
   # end
 
-  # @tag hub_id: :gossip_interval_test
-  # @tag sync_strategy: :gossip
-  # @tag listed_hooks: [
-  #        {Hook.cluster_join(), :local},
-  #        {Hook.registry_pid_inserted(), :global},
-  #        {Hook.registry_pid_removed(), :global}
-  #      ]
-  # test "gossip interval sync test", %{hub_id: hub_id} = context do
-  #   child_count = 100
-  #   child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
+  @tag hub_id: :gossip_interval_test
+  @tag sync_strategy: :gossip
+  @tag listed_hooks: [
+         {Hook.cluster_join(), :local},
+         {Hook.registry_pid_inserted(), :global},
+         {Hook.registry_pid_removed(), :global}
+       ]
+  test "gossip interval sync test", %{hub_id: hub_id} = context do
+    child_count = 100
+    child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
 
-  #   # Locally start children without propagating to the rest of the cluster.
-  #   Common.periodic_sync_base(context, child_specs, :add)
+    # Locally start children without propagating to the rest of the cluster.
+    Common.periodic_sync_base(context, child_specs, :add)
 
-  #   # TODO: this is failing sometimes
-  #   # Manually trigger the periodic sync.
-  #   Common.trigger_periodc_sync(context, child_specs, :add)
+    # Manually trigger the periodic sync.
+    Common.trigger_periodc_sync(context, child_specs, :add)
 
-  #   # TODO: this is failing sometimes
-  #   # Test if data is synchronized in the cluster.
-  #   Common.validate_sync(context)
+    # Test if data is synchronized in the cluster.
+    Common.validate_sync(context)
 
-  #   # Locally stop children without propagating to the rest of the cluster.
-  #   Common.periodic_sync_base(context, child_specs, :rem)
+    # Locally stop children without propagating to the rest of the cluster.
+    Common.periodic_sync_base(context, child_specs, :rem)
 
-  #   # Manually trigger the periodic sync.
-  #   Common.trigger_periodc_sync(context, child_specs, :rem)
+    # Manually trigger the periodic sync.
+    Common.trigger_periodc_sync(context, child_specs, :rem)
 
-  #   # Test if data is synchronized in the cluster.
-  #   Common.validate_sync(context)
-  # end
+    # Test if data is synchronized in the cluster.
+    Common.validate_sync(context)
+  end
 
   # @tag redun_strategy: :replication
   # @tag hub_id: :redunc_activ_pass_test
@@ -173,66 +176,66 @@ defmodule Test.IntegrationTest do
   #   Common.validate_redundancy_mode(context)
   # end
 
-  @tag redun_strategy: :singularity
-  @tag hub_id: :redunc_singulary_test
-  @tag listed_hooks: [
-         {Hook.cluster_join(), :local},
-         {Hook.registry_pid_inserted(), :local}
-       ]
-  test "redundancy with singularity", %{hub_id: hub_id} = context do
-    child_count = 100
-    child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
+  # @tag redun_strategy: :singularity
+  # @tag hub_id: :redunc_singulary_test
+  # @tag listed_hooks: [
+  #        {Hook.cluster_join(), :local},
+  #        {Hook.registry_pid_inserted(), :local}
+  #      ]
+  # test "redundancy with singularity", %{hub_id: hub_id} = context do
+  #   child_count = 100
+  #   child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
 
-    # Starts children on all nodes.
-    Common.sync_base_test(context, child_specs, :add)
+  #   # Starts children on all nodes.
+  #   Common.sync_base_test(context, child_specs, :add)
 
-    # Tests if all child_specs are used for starting children.
-    Common.validate_registry_length(context, child_specs)
+  #   # Tests if all child_specs are used for starting children.
+  #   Common.validate_registry_length(context, child_specs)
 
-    # Tests if all child_specs are started on all nodes.
-    Common.validate_started_children(context, child_specs)
+  #   # Tests if all child_specs are started on all nodes.
+  #   Common.validate_started_children(context, child_specs)
 
-    # Tests redundancy and check if started children's count matches replication factor.
-    Common.validate_singularity(context)
-  end
+  #   # Tests redundancy and check if started children's count matches replication factor.
+  #   Common.validate_singularity(context)
+  # end
 
-  @tag hub_id: :divergence_test
-  @tag partition_strategy: :div
-  @tag listed_hooks: [
-         {Hook.cluster_join(), :local},
-         {Hook.post_nodes_redistribution(), :global}
-       ]
-  test "partition divergence test",
-       %{hub_id: hub_id, peer_nodes: peers, listed_hooks: lh} = context do
-    :net_kernel.monitor_nodes(true)
+  # @tag hub_id: :divergence_test
+  # @tag partition_strategy: :div
+  # @tag listed_hooks: [
+  #        {Hook.cluster_join(), :local},
+  #        {Hook.post_nodes_redistribution(), :global}
+  #      ]
+  # test "partition divergence test",
+  #      %{hub_id: hub_id, peer_nodes: peers, listed_hooks: lh} = context do
+  #   :net_kernel.monitor_nodes(true)
 
-    peer_to_start = 5
+  #   peer_to_start = 5
 
-    new_peers =
-      TestNode.start_nodes(
-        peer_to_start,
-        prefix: :divergence_test
-      )
+  #   new_peers =
+  #     TestNode.start_nodes(
+  #       peer_to_start,
+  #       prefix: :divergence_test
+  #     )
 
-    peer_names = for {peer, pid} <- new_peers, do: peer
+  #   peer_names = for {peer, pid} <- new_peers, do: peer
 
-    Bootstrap.gen_hub(context)
-    |> Bootstrap.start_hubs(peer_names, lh, true)
+  #   Bootstrap.gen_hub(context)
+  #   |> Bootstrap.start_hubs(peer_names, lh, true)
 
-    assert ProcessHub.is_partitioned?(hub_id) === false
+  #   assert ProcessHub.is_partitioned?(hub_id) === false
 
-    end_num = peer_to_start - 1
+  #   end_num = peer_to_start - 1
 
-    Enum.reduce(peer_to_start..end_num, new_peers, fn numb, acc ->
-      removed_peers = Common.stop_peers(acc, 1)
-      Bag.receive_multiple(numb, Hook.post_nodes_redistribution())
-      Enum.filter(acc, fn node -> !Enum.member?(removed_peers, node) end)
-    end)
+  #   Enum.reduce(peer_to_start..end_num, new_peers, fn numb, acc ->
+  #     removed_peers = Common.stop_peers(acc, 1)
+  #     Bag.receive_multiple(numb, Hook.post_nodes_redistribution())
+  #     Enum.filter(acc, fn node -> !Enum.member?(removed_peers, node) end)
+  #   end)
 
-    assert ProcessHub.is_partitioned?(hub_id) === false
+  #   assert ProcessHub.is_partitioned?(hub_id) === false
 
-    :net_kernel.monitor_nodes(false)
-  end
+  #   :net_kernel.monitor_nodes(false)
+  # end
 
   # @tag hub_id: :static_quroum_test
   # @tag partition_strategy: :static
@@ -284,193 +287,193 @@ defmodule Test.IntegrationTest do
   #   :net_kernel.monitor_nodes(false)
   # end
 
-  @tag hub_id: :dynamic_quroum_test
-  @tag partition_strategy: :dynamic
-  @tag quorum_size: 80
-  # 1 hour
-  @tag quorum_threshold_time: 3600
-  @tag listed_hooks: [
-         {Hook.cluster_join(), :local},
-         {Hook.cluster_leave(), :local},
-         # {Hook.priority_state_updated(), :local}
-         {Hook.post_nodes_redistribution(), :local}
-       ]
-  test "dynamic quorum with min of 70% of cluster",
-       %{hub_id: hub_id, peer_nodes: peers, listed_hooks: lh} = context do
-    :net_kernel.monitor_nodes(true)
+  # @tag hub_id: :dynamic_quroum_test
+  # @tag partition_strategy: :dynamic
+  # @tag quorum_size: 80
+  # # 1 hour
+  # @tag quorum_threshold_time: 3600
+  # @tag listed_hooks: [
+  #        {Hook.cluster_join(), :local},
+  #        {Hook.cluster_leave(), :local},
+  #        # {Hook.priority_state_updated(), :local}
+  #        {Hook.post_nodes_redistribution(), :local}
+  #      ]
+  # test "dynamic quorum with min of 70% of cluster",
+  #      %{hub_id: hub_id, peer_nodes: peers, listed_hooks: lh} = context do
+  #   :net_kernel.monitor_nodes(true)
 
-    Bag.receive_multiple(@nr_of_peers, Hook.post_nodes_redistribution())
+  #   Bag.receive_multiple(@nr_of_peers, Hook.post_nodes_redistribution())
 
-    assert ProcessHub.is_partitioned?(hub_id) === false
+  #   assert ProcessHub.is_partitioned?(hub_id) === false
 
-    peer_to_start = 4
-    new_peers = TestNode.start_nodes(peer_to_start, prefix: :dynamic_quorum_test)
-    peer_names = for {peer, pid} <- new_peers, do: peer
+  #   peer_to_start = 4
+  #   new_peers = TestNode.start_nodes(peer_to_start, prefix: :dynamic_quorum_test)
+  #   peer_names = for {peer, pid} <- new_peers, do: peer
 
-    Bootstrap.gen_hub(context) |> Bootstrap.start_hubs(peer_names, lh, true)
-    Bag.receive_multiple(peer_to_start, Hook.post_nodes_redistribution())
+  #   Bootstrap.gen_hub(context) |> Bootstrap.start_hubs(peer_names, lh, true)
+  #   Bag.receive_multiple(peer_to_start, Hook.post_nodes_redistribution())
 
-    assert ProcessHub.is_partitioned?(hub_id) === false
+  #   assert ProcessHub.is_partitioned?(hub_id) === false
 
-    removed_peers = Common.stop_peers(new_peers, 1)
-    new_peers = Enum.filter(new_peers, fn node -> !Enum.member?(removed_peers, node) end)
-    Bag.receive_multiple(1, Hook.post_nodes_redistribution())
+  #   removed_peers = Common.stop_peers(new_peers, 1)
+  #   new_peers = Enum.filter(new_peers, fn node -> !Enum.member?(removed_peers, node) end)
+  #   Bag.receive_multiple(1, Hook.post_nodes_redistribution())
 
-    # At this point we still have 90% of cluster left.
-    assert ProcessHub.is_partitioned?(hub_id) === false
+  #   # At this point we still have 90% of cluster left.
+  #   assert ProcessHub.is_partitioned?(hub_id) === false
 
-    removed_peers = Common.stop_peers(new_peers, 1)
-    new_peers = Enum.filter(new_peers, fn node -> !Enum.member?(removed_peers, node) end)
-    Bag.receive_multiple(1, Hook.post_nodes_redistribution())
+  #   removed_peers = Common.stop_peers(new_peers, 1)
+  #   new_peers = Enum.filter(new_peers, fn node -> !Enum.member?(removed_peers, node) end)
+  #   Bag.receive_multiple(1, Hook.post_nodes_redistribution())
 
-    # At this point we still have 80% of cluster left.
-    assert ProcessHub.is_partitioned?(hub_id) === false
+  #   # At this point we still have 80% of cluster left.
+  #   assert ProcessHub.is_partitioned?(hub_id) === false
 
-    removed_peers = Common.stop_peers(new_peers, 1)
-    _new_peers = Enum.filter(new_peers, fn node -> !Enum.member?(removed_peers, node) end)
-    Bag.receive_multiple(1, Hook.post_nodes_redistribution())
+  #   removed_peers = Common.stop_peers(new_peers, 1)
+  #   _new_peers = Enum.filter(new_peers, fn node -> !Enum.member?(removed_peers, node) end)
+  #   Bag.receive_multiple(1, Hook.post_nodes_redistribution())
 
-    # At this point we have 70% of cluster left.
-    assert ProcessHub.is_partitioned?(hub_id) === true
+  #   # At this point we have 70% of cluster left.
+  #   assert ProcessHub.is_partitioned?(hub_id) === true
 
-    :net_kernel.monitor_nodes(false)
-  end
+  #   :net_kernel.monitor_nodes(false)
+  # end
 
-  @tag migr_strategy: :cold
-  @tag hub_id: :migration_coldswap_test
-  @tag redun_strategy: :replication
-  @tag replication_factor: 2
-  @tag listed_hooks: [
-         {Hook.cluster_join(), :local},
-         {Hook.cluster_leave(), :global},
-         {Hook.registry_pid_inserted(), :global},
-         {Hook.child_migrated(), :global}
-       ]
-  test "coldswap migration with replication",
-       %{hub_id: hub_id, replication_factor: rf, listed_hooks: lh} = context do
-    nodes_count = @nr_of_peers
-    child_count = 10
-    child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
+  # @tag migr_strategy: :cold
+  # @tag hub_id: :migration_coldswap_test
+  # @tag redun_strategy: :replication
+  # @tag replication_factor: 2
+  # @tag listed_hooks: [
+  #        {Hook.cluster_join(), :local},
+  #        {Hook.cluster_leave(), :global},
+  #        {Hook.registry_pid_inserted(), :global},
+  #        {Hook.child_migrated(), :global}
+  #      ]
+  # test "coldswap migration with replication",
+  #      %{hub_id: hub_id, replication_factor: rf, listed_hooks: lh} = context do
+  #   nodes_count = @nr_of_peers
+  #   child_count = 10
+  #   child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
 
-    # Stop hubs on peer nodes before we start.
-    Enum.each(Node.list(), fn node ->
-      :erpc.call(node, ProcessHub.Initializer, :stop, [hub_id])
-    end)
+  #   # Stop hubs on peer nodes before we start.
+  #   Enum.each(Node.list(), fn node ->
+  #     :erpc.call(node, ProcessHub.Initializer, :stop, [hub_id])
+  #   end)
 
-    # Confirm that hubs are stopped.
-    Bag.receive_multiple(nodes_count, Hook.cluster_leave())
+  #   # Confirm that hubs are stopped.
+  #   Bag.receive_multiple(nodes_count, Hook.cluster_leave())
 
-    # Starts children.
-    Common.sync_base_test(context, child_specs, :add)
+  #   # Starts children.
+  #   Common.sync_base_test(context, child_specs, :add)
 
-    # Add custom data to children.
-    Enum.each(child_specs, fn child_spec ->
-      {_child_spec, [{_, pid}]} = ProcessHub.child_lookup(hub_id, child_spec.id)
-      Process.send(pid, {:set_value, :handoff_data, child_spec.id}, [])
-    end)
+  #   # Add custom data to children.
+  #   Enum.each(child_specs, fn child_spec ->
+  #     {_child_spec, [{_, pid}]} = ProcessHub.child_lookup(hub_id, child_spec.id)
+  #     Process.send(pid, {:set_value, :handoff_data, child_spec.id}, [])
+  #   end)
 
-    # Restart hubs on peer nodes and confirm they are up and running.
-    Bootstrap.gen_hub(context)
-    |> Bootstrap.start_hubs(Node.list(), lh, true)
+  #   # Restart hubs on peer nodes and confirm they are up and running.
+  #   Bootstrap.gen_hub(context)
+  #   |> Bootstrap.start_hubs(Node.list(), lh, true)
 
-    ring = ProcessHub.Service.Ring.get_ring(hub_id)
-    local_node = node()
+  #   ring = ProcessHub.Service.Ring.get_ring(hub_id)
+  #   local_node = node()
 
-    # Get all children that have been migrated. Meaning the old ones are killed
-    # and spawned on other nodes. We can check all that no longer live on the main
-    # node.
-    migrated_children =
-      Enum.map(child_specs, fn child_spec ->
-        {child_spec.id, ProcessHub.Service.Ring.key_to_nodes(ring, child_spec.id, rf)}
-      end)
-      |> Enum.filter(fn {_, nodes} ->
-        !Enum.member?(nodes, local_node)
-      end)
+  #   # Get all children that have been migrated. Meaning the old ones are killed
+  #   # and spawned on other nodes. We can check all that no longer live on the main
+  #   # node.
+  #   migrated_children =
+  #     Enum.map(child_specs, fn child_spec ->
+  #       {child_spec.id, ProcessHub.Service.Ring.key_to_nodes(ring, child_spec.id, rf)}
+  #     end)
+  #     |> Enum.filter(fn {_, nodes} ->
+  #       !Enum.member?(nodes, local_node)
+  #     end)
 
-    # Confirm that all migrated children have been updated.
-    Bag.receive_multiple(
-      length(migrated_children),
-      Hook.registry_pid_inserted(),
-      error_msg: "Child added timeout"
-    )
+  #   # Confirm that all migrated children have been updated.
+  #   Bag.receive_multiple(
+  #     length(migrated_children),
+  #     Hook.registry_pid_inserted(),
+  #     error_msg: "Child added timeout"
+  #   )
 
-    Bag.receive_multiple(
-      length(migrated_children) * rf,
-      Hook.registry_pid_inserted(),
-      error_msg: "Child added timeout"
-    )
-  end
+  #   Bag.receive_multiple(
+  #     length(migrated_children) * rf,
+  #     Hook.registry_pid_inserted(),
+  #     error_msg: "Child added timeout"
+  #   )
+  # end
 
-  @tag migr_strategy: :hot
-  @tag hub_id: :migration_hotswap_test
-  @tag migr_handover: true
-  @tag migr_retention: 1000
-  @tag listed_hooks: [
-         {Hook.cluster_join(), :local},
-         {Hook.cluster_leave(), :global},
-         {Hook.registry_pid_inserted(), :local},
-         {Hook.registry_pid_removed(), :local},
-         {Hook.post_nodes_redistribution(), :local}
-       ]
-  test "hotswap migration with handoff",
-       %{hub_id: hub_id, listed_hooks: lh} = context do
-    nodes_count = @nr_of_peers
-    child_count = 10
-    child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
+  # @tag migr_strategy: :hot
+  # @tag hub_id: :migration_hotswap_test
+  # @tag migr_handover: true
+  # @tag migr_retention: 1000
+  # @tag listed_hooks: [
+  #        {Hook.cluster_join(), :local},
+  #        {Hook.cluster_leave(), :global},
+  #        {Hook.registry_pid_inserted(), :local},
+  #        {Hook.registry_pid_removed(), :local},
+  #        {Hook.post_nodes_redistribution(), :local}
+  #      ]
+  # test "hotswap migration with handoff",
+  #      %{hub_id: hub_id, listed_hooks: lh} = context do
+  #   nodes_count = @nr_of_peers
+  #   child_count = 10
+  #   child_specs = Bag.gen_child_specs(child_count, Atom.to_string(hub_id))
 
-    # Stop hubs on peer nodes before we start.
-    Enum.each(Node.list(), fn node ->
-      :erpc.call(node, ProcessHub.Initializer, :stop, [hub_id])
-    end)
+  #   # Stop hubs on peer nodes before we start.
+  #   Enum.each(Node.list(), fn node ->
+  #     :erpc.call(node, ProcessHub.Initializer, :stop, [hub_id])
+  #   end)
 
-    # Confirm that hubs are stopped.
-    Bag.receive_multiple(nodes_count, Hook.cluster_leave(), error_msg: "Cluster leave timeout")
+  #   # Confirm that hubs are stopped.
+  #   Bag.receive_multiple(nodes_count, Hook.cluster_leave(), error_msg: "Cluster leave timeout")
 
-    # Starts children.
-    Common.sync_base_test(context, child_specs, :add)
+  #   # Starts children.
+  #   Common.sync_base_test(context, child_specs, :add)
 
-    # Add custom data to children.
-    Enum.each(child_specs, fn child_spec ->
-      {_child_spec, [{_, pid}]} = ProcessHub.child_lookup(hub_id, child_spec.id)
-      Process.send(pid, {:set_value, :handoff_data, child_spec.id}, [])
-    end)
+  #   # Add custom data to children.
+  #   Enum.each(child_specs, fn child_spec ->
+  #     {_child_spec, [{_, pid}]} = ProcessHub.child_lookup(hub_id, child_spec.id)
+  #     Process.send(pid, {:set_value, :handoff_data, child_spec.id}, [])
+  #   end)
 
-    # Restart hubs on peer nodes and confirm they are up and running.
-    Bootstrap.gen_hub(context)
-    |> Bootstrap.start_hubs(Node.list(), lh, true)
+  #   # Restart hubs on peer nodes and confirm they are up and running.
+  #   Bootstrap.gen_hub(context)
+  #   |> Bootstrap.start_hubs(Node.list(), lh, true)
 
-    Bag.receive_multiple(nodes_count, Hook.post_nodes_redistribution(),
-      error_msg: "Post redistribution timeout"
-    )
+  #   Bag.receive_multiple(nodes_count, Hook.post_nodes_redistribution(),
+  #     error_msg: "Post redistribution timeout"
+  #   )
 
-    ring = ProcessHub.Service.Ring.get_ring(hub_id)
-    local_node = node()
+  #   ring = ProcessHub.Service.Ring.get_ring(hub_id)
+  #   local_node = node()
 
-    # Get all children that have been migrated.
-    migrated_children =
-      Enum.map(child_specs, fn child_spec ->
-        {child_spec.id, ProcessHub.Service.Ring.key_to_nodes(ring, child_spec.id, 1)}
-      end)
-      |> Enum.filter(fn {_, nodes} ->
-        !Enum.member?(nodes, local_node)
-      end)
+  #   # Get all children that have been migrated.
+  #   migrated_children =
+  #     Enum.map(child_specs, fn child_spec ->
+  #       {child_spec.id, ProcessHub.Service.Ring.key_to_nodes(ring, child_spec.id, 1)}
+  #     end)
+  #     |> Enum.filter(fn {_, nodes} ->
+  #       !Enum.member?(nodes, local_node)
+  #     end)
 
-    mcl = length(migrated_children)
+  #   mcl = length(migrated_children)
 
-    # Confirm that all migrated children have been updated.
-    Bag.receive_multiple(mcl, Hook.registry_pid_removed())
+  #   # Confirm that all migrated children have been updated.
+  #   Bag.receive_multiple(mcl, Hook.registry_pid_removed())
 
-    # Validate the data.
-    Enum.each(migrated_children, fn {child_id, _nodes} ->
-      pid =
-        ProcessHub.child_lookup(hub_id, child_id)
-        |> elem(1)
-        |> Enum.at(0)
-        |> elem(1)
+  #   # Validate the data.
+  #   Enum.each(migrated_children, fn {child_id, _nodes} ->
+  #     pid =
+  #       ProcessHub.child_lookup(hub_id, child_id)
+  #       |> elem(1)
+  #       |> Enum.at(0)
+  #       |> elem(1)
 
-      handover_data = GenServer.call(pid, {:get_value, :handoff_data})
+  #     handover_data = GenServer.call(pid, {:get_value, :handoff_data})
 
-      assert handover_data === child_id, "Data mismatch for child #{child_id}"
-    end)
-  end
+  #     assert handover_data === child_id, "Data mismatch for child #{child_id}"
+  #   end)
+  # end
 end